requirements.txt
# Spotify API
spotipy==2.23.0

# Data manipulation and analysis
pandas==2.2.1

# Environment variables
python-dotenv==1.0.1

# Database interaction
SQLAlchemy==2.0.28

# PostgreSQL adapter for Python
psycopg2-binary==2.9.9
